fix hide\show book title in scroll mode\book mode

+++ app/src/main/java/com/foobnix/ui2/AppDB.java
@@ -226,7 +226,9 @@ public class AppDB {
     // }
 
     public void deleteAllData() {
-        fileMetaDao.deleteAll();
+        if(fileMetaDao!=null) {
+            fileMetaDao.deleteAll();
+        }
     }
 
     public List<FileMeta> deleteAllSafe() {
